_System/scripts/memory/requirements-memory.txt
# PARA Vault Memory Index - Phase A dependencies
# Install: pip install -r requirements-memory.txt

# Core ML
sentence-transformers==5.2.3
faiss-cpu==1.13.2
numpy==2.4.2
torch==2.10.0
transformers==5.2.0

# Supporting
PyYAML==6.0.3
scikit-learn==1.8.0
scipy==1.17.1

# Phase B: Extraction
PyMuPDF==1.27.1
ebooklib==0.20
python-docx==1.2.0
beautifulsoup4==4.14.3
lxml==6.0.2
